+++ packages/twenty-server/src/modules/messaging/crons/jobs/messaging-cron-job.module.ts
@@ -7,7 +7,7 @@ import { DataSourceEntity } from 'src/engine/metadata-modules/data-source/data-s
 import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repository/object-metadata-repository.module';
 import { GmailMessagesImportCronJob } from 'src/modules/messaging/crons/jobs/gmail-messages-import.cron.job';
 import { GmailPartialMessageListFetchCronJob } from 'src/modules/messaging/crons/jobs/gmail-partial-message-list-fetch.cron.job';
-import { GmailMessagesImportModule } from 'src/modules/messaging/services/gmail-fetch-message-content-from-cache/gmail-messages-import.module';
+import { GmailMessagesImportModule } from 'src/modules/messaging/services/gmail-messages-import/gmail-messages-import.module';
 import { MessageChannelWorkspaceEntity } from 'src/modules/messaging/standard-objects/message-channel.workspace-entity';
 
 @Module({
