+++ packages/twenty-server/src/modules/messaging/jobs/gmail-full-message-list-fetch.job.ts
@@ -3,7 +3,7 @@ import { Injectable, Logger } from '@nestjs/common';
 import { MessageQueueJob } from 'src/engine/integrations/message-queue/interfaces/message-queue-job.interface';
 
 import { GoogleAPIRefreshAccessTokenService } from 'src/modules/connected-account/services/google-api-refresh-access-token/google-api-refresh-access-token.service';
-import { GmailFullMessageListFetchService } from 'src/modules/messaging/services/gmail-full-sync/gmail-full-message-list-fetch.service';
+import { GmailFullMessageListFetchService } from 'src/modules/messaging/services/gmail-full-message-list-fetch/gmail-full-message-list-fetch.service';
 
 export type GmailFullMessageListFetchJobData = {
   workspaceId: string;
