+++ packages/twenty-server/src/modules/messaging/jobs/messaging-job.module.ts
@@ -13,8 +13,8 @@ import { DeleteConnectedAccountAssociatedMessagingDataJob } from 'src/modules/me
 import { GmailFullMessageListFetchJob } from 'src/modules/messaging/jobs/gmail-full-message-list-fetch.job';
 import { GmailPartialMessageListFetchJob } from 'src/modules/messaging/jobs/gmail-partial-message-list-fetch.job';
 import { MessagingCreateCompanyAndContactAfterSyncJob } from 'src/modules/messaging/jobs/messaging-create-company-and-contact-after-sync.job';
-import { GmailFullMessageListFetchModule } from 'src/modules/messaging/services/gmail-full-sync/gmail-full-message-list-fetch.module';
-import { GmailPartialMessageListFetchModule } from 'src/modules/messaging/services/gmail-partial-sync/gmail-partial-message-list-fetch.module';
+import { GmailFullMessageListFetchModule } from 'src/modules/messaging/services/gmail-full-message-list-fetch/gmail-full-message-list-fetch.module';
+import { GmailPartialMessageListFetchModule } from 'src/modules/messaging/services/gmail-partial-message-list-fetch/gmail-partial-message-list-fetch.module';
 import { ThreadCleanerModule } from 'src/modules/messaging/services/thread-cleaner/thread-cleaner.module';
 import { MessageChannelMessageAssociationWorkspaceEntity } from 'src/modules/messaging/standard-objects/message-channel-message-association.workspace-entity';
 import { MessageChannelWorkspaceEntity } from 'src/modules/messaging/standard-objects/message-channel.workspace-entity';
